Changed some settings of wireless
I gave the wireless thingy a boost of its range. I changed the base range to 32 (Was 16) and the boosters to +2 blocks each instead of one, and I also changed the wireless booster's stack size to 64 instead of 16.

These are fair because the wireless access point is underpowered for it's price, so are the boosters because they cost a certus quartz to make and certus quartz are quite rare. The change to stack size shouldn't be a balance issue because it just allows you to add more boosters into the access point.